condvar-stl: replace pthread_cond calls with std::condition_variable
Replace the explicit libpthread dependence of the previous
implementation with one based on std::condition_variable and
std::chrono.  The STL-based implementation is shorter and simpler, but
the function that times out is much slower.  That may be because the
pthread version calls mutex.try_lock() and returns upon failure, while
the STL version calls the std::unique_lock ctor without
std::defer_lock, which means that it blocks upon failure.  strace and
ltrace demonstrate that the STL version is mostly syntactic sugar over
libthread.