condvar-pthread: eliminate race condition by turning off UBSAN
The run_many_condvar-pthread_tests.sh script fails about every 1/100 runs with
the Makefile setting both UBSAN and TSAN in the compilation, as in the previous
commit.  With only 2 threads, it fails about 1/400 times.  The test also runs
about 400x without failing if the two Parity::odd threads are started first so
that they must block and wait for the Parity::even ones.  With four threads
ordered in the same way as before, but with UBSAN turned off and TSAN left on,
the test will run 10k times without failing.