condvar-rtpi: adapt STL version by signalling while holding lock
The realtime kernel's priority-inheritance mechanism reuses the
mechanism on the kernel's lockdep validator.  Fundamentally, userspace
must somehow make the futex() system calls which underlie mutexes in
order to make a lock visible to the kernel.  The realtime scheduler
can only reason about which high-priority userspace threads are
waiting on a lock and promote the holder if the lock is locked.  Thus
the condition variable implementation in librtpi needs a lock
parameter for the notify_one() and notify_all() functions.  In
contrast, the STL and libpthread versions of the code have the option
to release the lock before signaling.

TSAN shows that both test cases have data races, which I will address
later.